+++ R/plot_covid19_stripes.R
@@ -9,9 +9,9 @@
 #' @param data The data frame to base the plot on. Should be a merged data
 #'     frame obtained by \link{download_merged_data} and defaults to
 #'     \code{download_merged_data(cached = TRUE, silent = TRUE)}.
-#' @param type The statistic that you want to plot. Needs to be either "confirmed",
-#'     "deaths", "revovered" or "active", defined as the difference of "confirmed"
-#'     and "recovered".
+#' @param type The statistic that you want to plot. Needs to be either
+#'     "confirmed", "deaths", "recovered" or "active", defined as the difference
+#'     of "confirmed", "deaths" and "recovered".
 #' @param min_cases Only countries that have a maximum of \code{type} cases higher
 #'     than \code{min_cases} during the data period are included in the plot.
 #'     Uses reasonable defaults depending on \code{type} and
@@ -101,7 +101,7 @@ plot_covid19_stripes <- function(
 
   data <- data %>%
     dplyr::mutate(
-      active = .data$confirmed - .data$recovered,
+      active = .data$confirmed - .data$recovered -.data$deaths,
       orig_type = !! rlang::sym(type)
     )